+++ cli/commands/toolchain/index.ts
@@ -22,6 +22,7 @@ import * as wasmtime from "./wasmtime.js";
 import * as lintoko from "./lintoko.js";
 import * as wasmOpt from "./wasm-opt.js";
 import { FILE_PATH_REGEX } from "../../constants.js";
+import { getPocketIcUrl } from "../../helpers/pocket-ic-startup.js";
 import * as toolchainUtils from "./toolchain-utils.js";
 import type { ReleaseInfo } from "./release-tags.js";
 import { normalizeBinaryenVersion } from "../../helpers/binaryen-version.js";
@@ -240,7 +241,7 @@ async function installAll({ silent = false, verbose = false } = {}) {
   if (config.toolchain?.wasmtime) {
     await download("wasmtime", config.toolchain.wasmtime, { silent, verbose });
   }
-  if (config.toolchain?.["pocket-ic"]) {
+  if (config.toolchain?.["pocket-ic"] && !getPocketIcUrl()) {
     await download("pocket-ic", config.toolchain["pocket-ic"], {
       silent,
       verbose,